WFLY-7617 Remove workaround for WFCORE-1941 in infinispan subsystem transform/reject tests
